+++ RELEASE_NOTES.md
@@ -3,3 +3,4 @@
 **Fixes**
 
 * Fixed an issue where images created with the VRC Camera could show incorrect world data.
+* Fxied an issue where webhook urls/names aren't loaded correctly in Media Relay.